consumer/requirements.txt
boto3==1.29.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
moto==4.1.14 